Added pastEvents to loader
